# aegisflow-openapi-ibm.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AegisFlow",
    "description": "AegisFlow - AI-Powered Autonomous Supply Chain Decision Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/logistics/{shipment_id}": {
      "get": {
        "tags": [
          "Logistics Management"
        ],
        "summary": "Get Logistics Item",
        "operationId": "get_logistics_item_api_logistics__shipment_id__get",
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inventory/{item_id}": {
      "get": {
        "tags": [
          "Inventory Management"
        ],
        "summary": "Get Inventory Item",
        "operationId": "get_inventory_item_api_inventory__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/risk-analysis": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Analyze Supply Chain Risk",
        "description": "Executes real IBM watsonx.ai Granite model text generation\nfor supply chain risk analysis.\n\nFlow:\nInput -> FastAPI -> IBM watsonx.ai Granite\n-> Structured Risk Analysis JSON.",
        "operationId": "analyze_supply_chain_risk_api_ai_risk_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskAnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/chat": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Chat With Ai",
        "description": "Existing AegisFlow Supply-Chain AI Chatbot.\n\nThis endpoint is intentionally preserved.",
        "operationId": "chat_with_ai_api_ai_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/orchestrate-chat": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Chat With Watson Orchestrate",
        "description": "Sends a user message to the deployed\nIBM watsonx Orchestrate AegisFlow agent.\n\nThis is a separate endpoint so the existing\n/chat endpoint is not affected.",
        "operationId": "chat_with_watson_orchestrate_api_ai_orchestrate_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/weather": {
      "get": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Get Live Location Weather",
        "description": "Retrieves live current weather observations for any\nlocation via Location and Weather Agents.",
        "operationId": "get_live_location_weather_api_ai_weather_get",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Location name or city",
              "title": "Location"
            },
            "description": "Location name or city"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/logistics/{shipment_id}": {
      "get": {
        "tags": [
          "Logistics Management"
        ],
        "summary": "Get Logistics Item",
        "operationId": "get_logistics_item_api_v1_logistics__shipment_id__get",
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inventory/{item_id}": {
      "get": {
        "tags": [
          "Inventory Management"
        ],
        "summary": "Get Inventory Item",
        "operationId": "get_inventory_item_api_v1_inventory__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ai/risk-analysis": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Analyze Supply Chain Risk",
        "description": "Executes real IBM watsonx.ai Granite model text generation\nfor supply chain risk analysis.\n\nFlow:\nInput -> FastAPI -> IBM watsonx.ai Granite\n-> Structured Risk Analysis JSON.",
        "operationId": "analyze_supply_chain_risk_api_v1_ai_risk_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskAnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/chat": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Chat With Ai",
        "description": "Existing AegisFlow Supply-Chain AI Chatbot.\n\nThis endpoint is intentionally preserved.",
        "operationId": "chat_with_ai_api_v1_ai_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/orchestrate-chat": {
      "post": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Chat With Watson Orchestrate",
        "description": "Sends a user message to the deployed\nIBM watsonx Orchestrate AegisFlow agent.\n\nThis is a separate endpoint so the existing\n/chat endpoint is not affected.",
        "operationId": "chat_with_watson_orchestrate_api_v1_ai_orchestrate_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/weather": {
      "get": {
        "tags": [
          "IBM watsonx.ai Granite Intelligence"
        ],
        "summary": "Get Live Location Weather",
        "description": "Retrieves live current weather observations for any\nlocation via Location and Weather Agents.",
        "operationId": "get_live_location_weather_api_v1_ai_weather_get",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Location name or city",
              "title": "Location"
            },
            "description": "Location name or city"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatInput": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "User query for supply chain intelligence"
          },
          "conversation_history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation History",
            "description": "Previous messages in thread"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatInput"
      },
      "ChatOutput": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "tools_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools Used"
          },
          "tool_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Data"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "response",
          "tools_used",
          "model"
        ],
        "title": "ChatOutput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name",
            "default": "Default Item"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku",
            "default": "SKU-GENERIC"
          },
          "current_stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stock",
            "default": 1000
          },
          "daily_demand": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Demand",
            "default": 50
          },
          "reorder_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reorder Level",
            "default": 300
          },
          "days_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Remaining",
            "default": 20
          },
          "stockout_risk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockout Risk",
            "default": "Low"
          },
          "stockout_risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockout Risk Level",
            "default": "LOW"
          },
          "unit_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Cost",
            "default": 45.0
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "InventoryOut"
      },
...
      "RiskAnalysisInput": {
        "properties": {
          "supplier": {
            "type": "string",
            "title": "Supplier",
            "description": "Supplier name",
            "default": "ABC Electronics"
          },
          "shipment_delay_days": {
            "type": "integer",
            "title": "Shipment Delay Days",
            "description": "Days of shipment delay",
            "default": 4
          },
          "inventory_days": {
            "type": "integer",
            "title": "Inventory Days",
            "description": "Days of buffer inventory remaining",
            "default": 3
          },
          "weather_risk": {
            "type": "string",
            "title": "Weather Risk",
            "description": "Weather disruption factor",
            "default": "Severe rainfall"
          },
          "supplier_reliability": {
            "type": "number",
            "title": "Supplier Reliability",
            "description": "Supplier reliability score percentage",
            "default": 65.0
          }
        },
        "type": "object",
        "title": "RiskAnalysisInput"
      },
      "RiskAnalysisOutput": {
        "properties": {
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          }
        },
        "type": "object",
        "required": [
          "risk_level",
          "risk_score",
          "reason",
          "recommendation",
          "confidence"
        ],
        "title": "RiskAnalysisOutput"
      },
...
      "ShipmentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "shipment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Id"
          },
          "shipment_code": {
            "type": "string",
            "title": "Shipment Code"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "delay_days": {
            "type": "integer",
            "title": "Delay Days"
          },
          "delay_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delay Category"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "cargo_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo Description"
          },
          "carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier"
          },
          "origin_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lat"
          },
          "origin_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Lng"
          },
          "dest_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Lat"
          },
          "dest_lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Lng"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shipment_code",
          "status",
          "delay_days",
          "risk_level"
        ],
        "title": "ShipmentOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://submit-museum-katrina-going.trycloudflare.com"
    }
  ]
}